pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gcp-secret-manager-cli"
version = "0.0.2"
description = "A command-line tool for managing GCP Secret Manager"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [{ name = "Albert Liu", email = "dreamerhyde@gmail.com" }]
keywords = ["gcp", "secret-manager", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0.0",
    "google-cloud-secret-manager>=2.0.0",
    "python-dotenv>=0.19.0",
    "rich>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/gcp-secret-manager-cli"
Documentation = "https://github.com/yourusername/gcp-secret-manager-cli#readme"
Repository = "https://github.com/yourusername/gcp-secret-manager-cli.git"

[project.scripts]
gcp-secrets = "gcp_secret_manager_cli.cli:cli"
sm = "gcp_secret_manager_cli.cli:cli"

[tool.black]
line-length = 80
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3
